+++ src/main/java/org/elasticsearch/indices/breaker/CircuitBreakerService.java
@@ -28,7 +28,7 @@ import org.elasticsearch.common.settings.Settings;
  * Interface for Circuit Breaker services, which provide breakers to classes
  * that load field data.
  */
-public abstract class CircuitBreakerService extends AbstractLifecycleComponent<InternalCircuitBreakerService> {
+public abstract class CircuitBreakerService extends AbstractLifecycleComponent<CircuitBreakerService> {
 
     protected CircuitBreakerService(Settings settings) {
         super(settings);
